MevzuatSistemi/requirements.txt
# Mevzuat Belge Analiz & Sorgulama Sistemi - Gereksinimler

# Temel GUI Framework
PyQt5>=5.15.10

# Konfigürasyon
PyYAML>=6.0.1

# Veritabanı (built-in Python)
# sqlite3

# Belge İşleme
PyMuPDF>=1.23.10  # fitz
pdfplumber>=0.10.3
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic>=0.4.27

# OCR Desteği
pytesseract>=0.3.10
Pillow>=10.0.1

# Semantik Arama ve ML
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.4
scikit-learn>=1.3.2
torch>=2.1.0  # sentence-transformers için
transformers>=4.35.0

# Metin İşleme
nltk>=3.8.1
regex>=2023.10.3

# Dosya İzleme
watchdog>=3.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.1
psutil>=5.9.6
chardet>=5.2.0

# Web ve HTTP
requests>=2.31.0
beautifulsoup4>=4.12.2

# Rapor Oluşturma
reportlab>=4.0.6
jinja2>=3.1.2

# Grafik ve Visualization (opsiyonel)
matplotlib>=3.7.4
seaborn>=0.12.2

# EXE Packaging
pyinstaller>=6.2.0

# Windows Desteği
pywin32>=306

# Development ve Test
pytest>=7.4.3
pytest-qt>=4.2.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-mock>=3.11.1
coverage>=7.3.0
